+++ ledger/sandbox-classic/BUILD.bazel
@@ -107,6 +107,8 @@ da_scala_test_suite(
     ],
     deps = [
         ":sandbox-classic",
+        "//bazel_tools/runfiles:scala_runfiles",
+        "//daml-lf/archive:daml_lf_archive_reader",
         "//daml-lf/data",
         "//ledger/caching",
         "//ledger/cli-opts",
@@ -114,10 +116,13 @@ da_scala_test_suite(
         "//ledger/ledger-api-common:ledger-api-common-scala-tests-lib",
         "//ledger/ledger-api-domain",
         "//ledger/ledger-configuration",
+        "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
+        "//ledger/test-common:dar-files-default-lib",
         "//libs-scala/ports",
+        "@maven//:io_netty_netty_handler",
     ],
 )
