--- ledger/sandbox-common/src/main/scala/platform/sandbox/cli/SandboxCli.scala
@@ -1,18 +1,0 @@
-// Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.platform.sandbox.cli
-
-import com.daml.platform.sandbox.config.SandboxConfig
-import scopt.OptionParser
-
-trait SandboxCli {
-
-  def defaultConfig: SandboxConfig
-
-  protected def parser: OptionParser[SandboxConfig]
-
-  def parse(args: Array[String]): Option[SandboxConfig] =
-    parser.parse(args, defaultConfig)
-
-}
